Add basic editing and saving infrastructure
This renders an (empty) canvas element and allows
to change the publication flag of the pattern currently
edited.

This prepares everything for more elaborate editing,
because now we have javascript working and an API
set up for saving changes without reloading the page.
Next up would be rendering the pattern in javascript
in the canvas element. Then, when we implement changes
to the pattern, we need to store these using the API
functionality (but only when explicitely triggered,
locally, all changes need to be stored in the localStorage
every few seconds to prevent data loss due to page
refresh or browser restart).