Fix verification, add cli tools, create primary group on user creation
Outlook mail addresses use link protection which results in multiple
calls to the verification url. Now, we accept this, i.e. if already
verified, we accept further verifications as valid.

When creating a user, the associated primary group is created as well.
Without this, it is not possible to create patterns.

Some basic CLI tools are implemented for listing, displaying and
deleting users. Further such tools will come later on.